+++ module/nuls-consensus/nuls-consensus-poc/src/main/java/io/nuls/poc/pbft/model/VoteData.java
@@ -8,7 +8,7 @@ import io.nuls.base.data.NulsHash;
 public class VoteData {
     private long height;
 
-    private int round;
+    private long round;
 
     private NulsHash hash;
 
@@ -32,11 +32,11 @@ public class VoteData {
         this.height = height;
     }
 
-    public int getRound() {
+    public long getRound() {
         return round;
     }
 
-    public void setRound(int round) {
+    public void setRound(long round) {
         this.round = round;
     }
